Update createtables.py
modified data types

+++ pythonlab/createtables.py
@@ -13,9 +13,9 @@ def create_tables():
     command2 = """CREATE TABLE cities (
                 city_name VARCHAR(255) NOT NULL,
                 state_name VARCHAR(255) NOT NULL,
-                city_pop int(255) NOT NULL,
-                city_lat real(255) NOT NULL,
-                city_lon real(255) NOT NULL
+                city_pop NUMERIC NOT NULL,
+                city_lat float(255) NOT NULL,
+                city_lon float(255) NOT NULL
                 )"""
 
     conn = None
